Working minified messd-up

+++ src/op/edge.h
@@ -8,11 +8,11 @@ extern "C" {
 #endif
 
 typedef struct {
-    uint16_t _last;
+    double _last;
 } t_edge;
 
 void edge_init(t_edge *self);
-void edge_process(t_edge *self, uint16_t *in, uint16_t *out);
+void edge_process(t_edge *self, double *in, double *out);
 
 #ifdef __cplusplus
 }
